raftNode.id: convert type to `uint64`
This is more consistent with the types used elsewhere for IDs.

+++ contrib/raftexample/main.go
@@ -23,7 +23,7 @@ import (
 
 func main() {
 	cluster := flag.String("cluster", "http://127.0.0.1:9021", "comma separated cluster peers")
-	id := flag.Int("id", 1, "node ID")
+	id := flag.Uint64("id", 1, "node ID")
 	kvport := flag.Int("port", 9121, "key-value server port")
 	join := flag.Bool("join", false, "join an existing cluster")
 	flag.Parse()
